test: shard CameraController Harness tests
